Factor HTTP handlers etc., into a driver object
Among other things, this

 * puts the state in a struct, instead of global vars
 * removes clutter from main (by sweeping under a subpackage)
